Create 20. Valid Parentheses
Approach:
1. Initialize a stack
2. Iterate Over the String
3. Final Check

Time complexity:
O(n), where n is the length of the string. Each character is processed once and each push/pop operation on the stack is O(1).

Space complexity:
O(n) in the worst case, where n is the length of the string. In the worst case, all characters could be opening brackets and thus all pushed onto the stack.