Create 769. Max Chunks To Make Sorted
**Intuition**
So it is an easy question. Let me sum up all the things I am doing here. So basically, there are two pointers, i and j. We can also say that the initial window.
There is one more pointer k which we use to travel no the window that is from i to j (included)
while travelling on the window using k if we get any number which is not lying in the range from i to j then we will increase the size of the window that is j++ until the kth integer don't lie in the window .
At last, if all elements in a particular window that is from i to j lie in the window we will increase the count++ (which tells that we have got a valid window )
And after that, we will move to the next part of the array starting everything from scratch

**Complexity**
Time complexity: O(n)
Space complexity:O(1)